28/08

+++ src/main/java/ru/geekbrains/Fruit.java
@@ -1,5 +1,6 @@
 package ru.geekbrains;
 
 public abstract class Fruit {
-    int weight;
+    double weight;
+
 }
